requirements.txt
certifi==2020.4.5.1
asgiref==3.2.7
certifi==2020.4.5.1
Django==3.0.6
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1

django-cors-headers==3.2.1

